Split the sitemap into a separate file
