+++ package.json
@@ -23,6 +23,7 @@
   "homepage": "https://github.com/NorthernLogic/polygigs#readme",
   "dependencies": {
     "gulp": "^3.9.1",
+    "gulp-prompt": "^0.2.0",
     "node-fetch": "^1.6.3",
     "sitemap": "^1.8.2"
   }
